refactor(server): decompose server.ts into Fastify plugins under src/routes/
Rebase onto latest origin/develop and re-apply the refactoring.

Extract route handlers from the monolithic 2449-line server.ts into
focused Fastify plugins under src/routes/:
- health.ts: health, metrics, alerts, swarm, handshake, audit
- auth.ts: key management, verification, SSE tokens, rotation
- sessions.ts: all session CRUD, interaction, hooks, SSE, batch
- events.ts: global SSE event stream
- pipelines.ts: pipeline CRUD
- templates.ts: session template CRUD
- diagnostics.ts: diagnostics, tools, webhooks, channels, latency

Also extract rate-limit utilities into src/rate-limit.ts to avoid
circular imports between server.ts and route plugins.

Incorporates all develop changes since branch divergence:
- AuditLogger + AlertManager initialization (#1418, #1419)
- Request ID correlation (#1416)
- /metrics auth (#1557)
- KEY_EXPIRED error code (#1403)
- claudeCommand RCE prevention (#1393)
- ?secret= redaction (#1393)
- Pipeline persistence (#1424)
- Removes consensus/model-router references (deleted in #1583)

server.ts reduced from 2449 to ~580 lines (bootstrap + middleware).
